+++ analyzer/analyzerparameters.h
@@ -419,7 +419,7 @@ public:
         m_analyzers << new AnalyzerParameters(idx++, "WilsonPro CAA", "100", "1500000", 240, 320, PREFIX_SERIAL_NUMBER_WILSON_PRO);
         m_analyzers << new AnalyzerParameters(idx++, "AA-1500 ZOOM SE", "100", "1500000", 480, 746, PREFIX_SERIAL_NUMBER_AA1500SE, "AA-1500SE");
         m_analyzers << new AnalyzerParameters(idx++, "AA-1500 SE", "100", "1500000", 480, 746, PREFIX_SERIAL_NUMBER_AA1500SE);
-        m_analyzers << new AnalyzerParameters(idx++, "Match", "100", "690000", 480, 480, PREFIX_SERIAL_NUMBER_MATCH);
+        m_analyzers << new AnalyzerParameters(idx++, "Match", "100", "750000", 480, 480, PREFIX_SERIAL_NUMBER_MATCH);
         m_analyzers << new AnalyzerParameters(idx++, "MATCH U", "100", "500000", 480, 480, PREFIX_SERIAL_NUMBER_MATCHU);
 #ifdef _DEBUG
         test();